Reorganize core instruments for Python
