+++ nautilus_core/model/src/instruments/crypto_future.rs
@@ -15,18 +15,12 @@
 
 #![allow(dead_code)] // Allow for development
 
-use std::{
-    collections::hash_map::DefaultHasher,
-    hash::{Hash, Hasher},
-};
+use std::hash::{Hash, Hasher};
 
 use anyhow::Result;
-use nautilus_core::{
-    python::{serialization::from_dict_pyo3, to_pyvalue_err},
-    time::UnixNanos,
-};
-use pyo3::{basic::CompareOp, prelude::*, types::PyDict};
-use rust_decimal::{prelude::ToPrimitive, Decimal};
+use nautilus_core::time::UnixNanos;
+use pyo3::prelude::*;
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 
 use super::Instrument;
@@ -219,131 +213,6 @@ impl Instrument for CryptoFuture {
 
     fn taker_fee(&self) -> Decimal {
         self.taker_fee
-    }
-}
-
-#[cfg(feature = "python")]
-#[pymethods]
-impl CryptoFuture {
-    #[allow(clippy::too_many_arguments)]
-    #[new]
-    fn py_new(
-        id: InstrumentId,
-        raw_symbol: Symbol,
-        underlying: Currency,
-        quote_currency: Currency,
-        settlement_currency: Currency,
-        expiration: UnixNanos,
-        price_precision: u8,
-        size_precision: u8,
-        price_increment: Price,
-        size_increment: Quantity,
-        margin_init: Decimal,
-        margin_maint: Decimal,
-        maker_fee: Decimal,
-        taker_fee: Decimal,
-        lot_size: Option<Quantity>,
-        max_quantity: Option<Quantity>,
-        min_quantity: Option<Quantity>,
-        max_notional: Option<Money>,
-        min_notional: Option<Money>,
-        max_price: Option<Price>,
-        min_price: Option<Price>,
-    ) -> PyResult<Self> {
-        Self::new(
-            id,
-            raw_symbol,
-            underlying,
-            quote_currency,
-            settlement_currency,
-            expiration,
-            price_precision,
-            size_precision,
-            price_increment,
-            size_increment,
-            margin_init,
-            margin_maint,
-            maker_fee,
-            taker_fee,
-            lot_size,
-            max_quantity,
-            min_quantity,
-            max_notional,
-            min_notional,
-            max_price,
-            min_price,
-        )
-        .map_err(to_pyvalue_err)
-    }
-
-    fn __hash__(&self) -> isize {
-        let mut hasher = DefaultHasher::new();
-        self.hash(&mut hasher);
-        hasher.finish() as isize
-    }
-
-    fn __richcmp__(&self, other: &Self, op: CompareOp, py: Python<'_>) -> Py<PyAny> {
-        match op {
-            CompareOp::Eq => self.eq(other).into_py(py),
-            _ => panic!("Not implemented"),
-        }
-    }
-
-    #[staticmethod]
-    #[pyo3(name = "from_dict")]
-    fn py_from_dict(py: Python<'_>, values: Py<PyDict>) -> PyResult<Self> {
-        from_dict_pyo3(py, values)
-    }
-    #[pyo3(name = "to_dict")]
-    fn py_to_dict(&self, py: Python<'_>) -> PyResult<PyObject> {
-        let dict = PyDict::new(py);
-        dict.set_item("type", stringify!(CryptoPerpetual))?;
-        dict.set_item("id", self.id.to_string())?;
-        dict.set_item("raw_symbol", self.raw_symbol.to_string())?;
-        dict.set_item("underlying", self.underlying.code.to_string())?;
-        dict.set_item("quote_currency", self.quote_currency.code.to_string())?;
-        dict.set_item(
-            "settlement_currency",
-            self.settlement_currency.code.to_string(),
-        )?;
-        dict.set_item("expiration", self.expiration.to_i64())?;
-        dict.set_item("price_precision", self.price_precision)?;
-        dict.set_item("size_precision", self.size_precision)?;
-        dict.set_item("price_increment", self.price_increment.to_string())?;
-        dict.set_item("size_increment", self.size_increment.to_string())?;
-        dict.set_item("margin_init", self.margin_init.to_f64())?;
-        dict.set_item("margin_maint", self.margin_maint.to_f64())?;
-        dict.set_item("maker_fee", self.margin_init.to_f64())?;
-        dict.set_item("taker_fee", self.margin_init.to_f64())?;
-        match self.lot_size {
-            Some(value) => dict.set_item("lot_size", value.to_string())?,
-            None => dict.set_item("lot_size", py.None())?,
-        }
-        match self.max_quantity {
-            Some(value) => dict.set_item("max_quantity", value.to_string())?,
-            None => dict.set_item("max_quantity", py.None())?,
-        }
-        match self.min_quantity {
-            Some(value) => dict.set_item("min_quantity", value.to_string())?,
-            None => dict.set_item("min_quantity", py.None())?,
-        }
-        match self.max_notional {
-            Some(value) => dict.set_item("max_notional", value.to_string())?,
-            None => dict.set_item("max_notional", py.None())?,
-        }
-        match self.min_notional {
-            Some(value) => dict.set_item("min_notional", value.to_string())?,
-            None => dict.set_item("min_notional", py.None())?,
-        }
-        match self.max_price {
-            Some(value) => dict.set_item("max_price", value.to_string())?,
-            None => dict.set_item("max_price", py.None())?,
-        }
-        match self.min_price {
-            Some(value) => dict.set_item("min_price", value.to_string())?,
-            None => dict.set_item("min_price", py.None())?,
-        }
-        Ok(dict.into())
     }
 }
